relocate thunks
